+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - delete reminders for form actions when the correct form is filled
  - add `email_activate_subject` and `email_reminder_subject` fields for configuration when sending emails
  - add an option to attach files when schedule a mail action for forms
+ - add language check `condition`
 
 # v5.1.5
 ### Bugfix